Unified the config logic of APIServer and ControllerMgr
