+++ go/cmd/apiserver/config.go
@@ -21,29 +21,12 @@ type (
 		Host string `yaml:"host"`
 		Port int    `yaml:"port"`
 	}
-	// K8sConfig is the configuration for k8s REST client.
-	K8sConfig struct {
-		QPS   float32 `yaml:"qps"`
-		Burst int     `yaml:"burst"`
-	}
 )
 
 // getK8sRestConfig parses the configuration file and returns the k8s REST client configuration
 // for Michelangelo API server.
-func getK8sRestConfig(provider config.Provider) (*rest.Config, error) {
-	flag.Parse()
-	conf, err := ctrl.GetConfig()
-	if err != nil {
-		return nil, err
-	}
-	k8sConfig := K8sConfig{}
-	err = provider.Get(k8sConfigKey).Populate(&k8sConfig)
-	if err != nil {
-		return nil, err
-	}
-	conf.QPS = k8sConfig.QPS
-	conf.Burst = k8sConfig.Burst
-	return conf, nil
+func getK8sRestConfig(provider config.Provider, configKey string) (*rest.Config, error) {
+	return baseconfig.GetK8sConfig(provider, configKey)
 }
 
 // getYARPCConfig parses the configuration file and returns the YARPC server configuration
@@ -55,7 +38,5 @@ func getYARPCConfig(provider config.Provider) (YARPCConfig, error) {
 }
 
 func getMetadataStorageConfig(provider config.Provider) (storage.MetadataStorageConfig, error) {
-	storageConfig := storage.MetadataStorageConfig{}
-	err := provider.Get(storageConfigKey).Populate(&storageConfig)
-	return storageConfig, err
+	return baseconfig.GetMetadataStorageConfig(provider, storageConfigKey)
 }
