+++ go/cmd/apiserver/config_test.go
@@ -9,68 +9,6 @@ import (
 	"go.uber.org/config"
 )
 
-var k8sConf = `
-apiVersion: v1
-clusters:
-- cluster:
-    certificate-authority-data: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSURMRENDQWhRQ0NRRFhhaHE1VHlvWVpUQU5CZ2txaGtpRzl3MEJBUXNGQURCWU1Rc3dDUVlEVlFRR0V3SlYKVXpFTk1Bc0dBMVVFQ0F3RVRtOXVaVEVOTUFzR0ExVUVCd3dFVG05dVpURU5NQXNHQTFVRUNnd0VUbTl1WlRFYwpNQm9HQTFVRUF3d1RiV2xqYUdWc1lXNW5aV3h2TFdzNGN5MWpZVEFlRncweU1UQXpNalF3TVRReE1qaGFGdzB6Ck1UQXpNakl3TVRReE1qaGFNRmd4Q3pBSkJnTlZCQVlUQWxWVE1RMHdDd1lEVlFRSURBUk9iMjVsTVEwd0N3WUQKVlFRSERBUk9iMjVsTVEwd0N3WURWUVFLREFST2IyNWxNUnd3R2dZRFZRUUREQk50YVdOb1pXeGhibWRsYkc4dAphemh6TFdOaE1JSUJJakFOQmdrcWhraUc5dzBCQVFFRkFBT0NBUThBTUlJQkNnS0NBUUVBeVlYQjU2NEhqZG8xClRZVERnTTVuNVJFSEU1am1hanNlR0QzNlNnUzdCazN5V0R3V0Z4T3VRaWNXRnFocnhpR2NScE16VndvZTNKQjAKRERjUTRLQzhLS3FSbC9oRWVybGE5UUpiU3gybEs4Ny9Fck5vaVh1OWNDeHdCcmlmbFN5WGQrWkhQa0pCdnd6Swp6cExxQXhqb00wMXM1K082SFRnZkoxWDBmeFordnRibC9EOGN0SUliV2t6K29NTUg0dmZNekVTVld1dDM2eTdYCkdCQjJOTndrSEtqMnR5eHNycjA5a1J5anRvcFFNN2ZGTnJHSDJtM3lrTmhJNzUxUWVsWWVreDRJdGxzTFFPR3IKUkxFTEkwL09CTTdWMFZPTC9OdlJBZndMTlRIUXVuUlpiVnB2RTZ3WHY1YlVtRDZLY2ErQ294LzQ1MlhnTFZOKwp5SWN5Mm9xQmVRSURBUUFCTUEwR0NTcUdTSWIzRFFFQkN3VUFBNElCQVFCSFBaNmlpVzNwdm1qL2t0RUUvYlRpCmVYL3U4SERRNktZdnRjYlByNEovVVlGQ01JZHdmVUtRclJrNnRkRUFReHhHcXhHOUlqL3B3RHpGYytUcjBXRHYKeTB3akx1WGw0K0tncEhUbFdWY1U1Q2QxanJlejFEUXpsZEhQaUpmRkhHQTVoNmVwdVRkbU9KREFmQnhRU2YzTAo4aUQwdU1Ma2NmUlYrVHk1S1BIcWdpYVBwclNScnRiUUhOZDFIcE5oOHpTaEJrK1N0amMyNUMwbEs4d0t4SUY2CnZBSW1kWDM1QmIrc1dzYjVEeFZrRTZtK05aT3dwU1V6Tk1NRnBQbXBlMEtMWlZlNDJDN0NzZ05YcWtUVVB4bWQKMms2REtySXA2aHg4YkNyVGVvakJBSlNnMGp3WCtYSnlnVVdXWFJra3F0dlNlYjhYOVJLeG40QXpvNHM5Sko4agotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0tCg==
-    server: https://10.1.130.1:1234
-  name: michelangelo
-contexts:
-- context:
-    cluster: michelangelo
-    user: system:michelangelo-apiserver
-  name: system:michelangelo-apiserver@michelangelo
-current-context: system:michelangelo-apiserver@michelangelo
-kind: Config
-preferences: {}
-users:
-- name: system:michelangelo-apiserver
-  user:
-    client-certificate-data: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSURNVENDQWhrQ0FRRXdEUVlKS29aSWh2Y05BUUVMQlFBd1dERUxNQWtHQTFVRUJoTUNWVk14RFRBTEJnTlYKQkFnTUJFNXZibVV4RFRBTEJnTlZCQWNNQkU1dmJtVXhEVEFMQmdOVkJBb01CRTV2Ym1VeEhEQWFCZ05WQkFNTQpFMjFwWTJobGJHRnVaMlZzYnkxck9ITXRZMkV3SGhjTk1qRXdNekkwTURFME1USTRXaGNOTXpFd016SXlNREUwCk1USTRXakJsTVFzd0NRWURWUVFHRXdKVlV6RU5NQXNHQTFVRUNBd0VUbTl1WlRFTk1Bc0dBMVVFQnd3RVRtOXUKWlRFWE1CVUdBMVVFQ2d3T2MzbHpkR1Z0T20xaGMzUmxjbk14SHpBZEJnTlZCQU1NRm0xcFkyaGxiR0Z1WjJWcwpieTFyT0hNdFlXUnRhVzR3Z2dFaU1BMEdDU3FHU0liM0RRRUJBUVVBQTRJQkR3QXdnZ0VLQW9JQkFRREY5bnAzCmt2S3FSQ3crZFV6K29sRmxBSEhzckM3eXhROWphRGhvSnhGaHovVkhPeGVNcXloQ1RXVGZXb0czMEkvb0poanQKT3NzK05tKzVnalA2eWZNRUNPNXhvTEVPNDg3cmtjL0ZDVzB2cHFkU0tGZlJSWVE4RUJRZUNvQ0dTbit4aEdsUgo0NEtUREdQRU5RRmtZNjh3czlHRTJqSmxEOGZCcURhRTJ2MlhlejRQNnlPQUdEYXNsZXBRWW44cVFhcG5nQmd4CmpZdGNWRTBjUG4xWXRnQkpSemdTZCszSVZQelV1UWZCSkVjcG1zWE9rUEVXYXNzVTR5em5qYnBrVE90RTYrKzkKUVd3Yld0QlFhU0doWGVkb2NIRG1yS25oVU4yWHpGTnZhNVFNcElkSWluY1M5anJUdGZ4SW5jTndDVnh2WjhNYgpabXgwaGtDRFBOc1M4L21GQWdNQkFBRXdEUVlKS29aSWh2Y05BUUVMQlFBRGdnRUJBRTl5cEo4ellLS3pKV0hUCkkwZytBYi9LaFdCOUJjcnUwZWxxU0pUdUdFL0NhcTJUNnJiU3h5T1hGTnZweGNBSXBvR04vcGhGeEM4VlVxMVMKSTlXZW56alIwSTFLdVY0ZlFsRDlqS3BNWWtQQXRCZHhQeEI2eEQxVlpIQ3diRlNpbjBxbmhaT0lvRUtmZGM4Wgp1SlBPMytqckh4WGFPMUpUeVVka1Vqb09talpmL1hsRitBZ3pzUm9FbGtObCs5TWxoRjUxK1pWQTloNXpScGZNCnlEVnl5dm5kOUd2N2tvWDBKdGZMV3Fyck5kZGhUSDRuakt1MU1sU3dMZVVnSk4wb3hnU1FEOWxqK3BRUUZXRkMKSlk5bTk3VjZreWVrelBkOHVxdS9tUDdwQy95Z2hoSFNaSDZjN3hyTE42eGR2UlJNRDl6K3d0RVBRd0IzbVpBTApBc08rK1k4PQotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0tCg==
-    client-key-data: LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlFcFFJQkFBS0NBUUVBeGZaNmQ1THlxa1FzUG5WTS9xSlJaUUJ4N0t3dThzVVBZMmc0YUNjUlljLzFSenNYCmpLc29RazFrMzFxQnQ5Q1A2Q1lZN1RyTFBqWnZ1WUl6K3NuekJBanVjYUN4RHVQTzY1SFB4UWx0TDZhblVpaFgKMFVXRVBCQVVIZ3FBaGtwL3NZUnBVZU9Da3d4anhEVUJaR092TUxQUmhOb3laUS9Id2FnMmhOcjlsM3MrRCtzagpnQmcyckpYcVVHSi9La0dxWjRBWU1ZMkxYRlJOSEQ1OVdMWUFTVWM0RW5mdHlGVDgxTGtId1NSSEtackZ6cER4CkZtckxGT01zNTQyNlpFenJST3Z2dlVGc0cxclFVR2tob1YzbmFIQnc1cXlwNFZEZGw4eFRiMnVVREtTSFNJcDMKRXZZNjA3WDhTSjNEY0FsY2IyZkRHMlpzZElaQWd6emJFdlA1aFFJREFRQUJBb0lCQVFDMWdCbWl2aWtjOE9vNwp6YUp0Si9kR2lsVHBjS2lPenh3ZndUTk9VdHQ5K1lQaDdneldYU3JKTjFwUDZoUmFxcTcyWXMraDJiRDBka2FUCktBeHVXSEFsOStaZWFiMU1jSk9McWlGN2Z2d0xuMWx6V3lmQTZHS29Ddm5XM1R3ZlkvaVZrbGhaeXlVeG16OXgKYWtPL2ZpSG1ua2R2N2ZvY2JJNXlRTDdUcWxDdGE2Vnl1amdDVTZOZ1RlUVQvUExiMjFpeVRZdHhTcWp6bjBFeApOZHA2UXB4R3k2bnRiUExUZGFBSVl1UTdLeFFsbzZjRlVsSElZWDROVm1LaUpTRTBNcVNaV3pCbS8wWDJrd3JsCmR3R0xHRFhIYWorR08vR283Wm1SWGJUTnpteXZDek9CSUJzOHlRRnBWQUdyYVBsYjZvczNLek5BZHFiYnZySWcKVnZRaW9XWFZBb0dCQU9kbjFxUnJuVFIrQVdDQ25CYmZXcHluNkluUXNyYnErZHhHYWRORXJuZ1V2ZExtYWx2YQorN2twR1hxMUhoSUFkTi9MRU1FQktOS2VYdDVGQnVLRnJqZGVMQWQ5WWJaNC9DYzFLZHY0L0srWnhONm44NElsCk5tOVRLdHovWHZUS0NSRmdHaEJoeEtPUVRYVDZ0Ri9WVWRQYjM5OEgwbmtzZmtGM0RobHRSbU12QW9HQkFOc0EKdDNsWTZmcDk5TTFyRktmN3hIR0NYaVdzV3IvLzhDMHJEOFYrZHU3YkQvc2d6VFNyL0pJdjJReEN1a1J4dmsxSQpwYU1YaWtjWmVvbTV3RzB3bUw1QkhaSWl2ZlZGQW5oWTVJQ1o1ajkveEExalhJbHlCRGtUZHFnamJHWnJRLzZpCldmYS9qeDNGRU9VQTVYWStsWktVeGVybk9PNmZlc3V6Tk16UHVoR0xBb0dCQUxEd0RnaFVuTFNwY0dZYUdEM0kKOU9ENTVtMlNYVVJPTVZVRHBpRTd6K2ZUZkQzSm55T3pNbXluQjJ0ekY1WU9NVTk1VnNzdEZzak0vWjhZeXFYawpMNHo0ZmRRUVErbWhZclNjQ3ZDKzFuOXlwVHpXMFBQL2ZqcnJMY2dqbjdpdXp2WXhORnk0VlFIMzhiSHpqSDRHCmYzWHVGcVRUdDFTZDk4QVl4M2didlFsVEFvR0FKcVNkdXovQktYNElNQ2J3NGlNK3FuakNmQXRKaUE5MUpjTXYKYVQzRFpxb295N3NoK21WT2o4ejVrM3hDdWNrSU4wTFdWMHpVRFcrbGU1L1hJRzB1eG9OZTRHWlk5bXBTNFVGdQpNSEwzZWNUbHB5Y2RNUE41WTBqWDZ4czFDVzFyOWdaWHNYNWpsbkVyWmYwZWdCclM4YVptdGVoTzEydzBrclR3CllDTlhSYmtDZ1lFQXdJSWRxWUtTOFdPdHZOMis1eGIxRkluTHozRTYvUENxMmI3SWVaNDY2T3NjVGZZOTJlYk4KRElHdlVKdFI1T29DeHJBSlBKcnBLVElUTFN5RlR6OExNekc2MzNSM0pId2dPYjlmWUswZE1FNDVselhVUHovNgplOFlZOGJ5Y1V1UllkcDJjNGovWDJkVTBEUnptQ1R2SWNRUi91MG5PUHdEN3gzaU0rSjhHeE44PQotLS0tLUVORCBSU0EgUFJJVkFURSBLRVktLS0tLQo=
-`
-
-func TestGetK8sConfig(t *testing.T) {
-	file, _ := os.CreateTemp(t.TempDir(), "k8s-conf")
-	os.Setenv("KUBECONFIG", file.Name())
-
-	yamlStr := `
-apiserver:
-  k8s:
-    qps: 300
-    burst: 600
-`
-	provider, err := config.NewYAML(config.Source(strings.NewReader(yamlStr)))
-	assert.NoError(t, err)
-
-	file.WriteString("invalid conf")
-	file.Sync()
-	conf, err := getK8sRestConfig(provider)
-	assert.Error(t, err) // error when the content of the configuration file is invalid
-
-	file.Seek(0, 0)
-	file.WriteString(k8sConf)
-	file.Close()
-
-	conf, err = getK8sRestConfig(provider)
-	assert.NoError(t, err)
-	assert.Equal(t, float32(300), conf.QPS)
-	assert.Equal(t, 600, conf.Burst)
-	assert.Equal(t, "https://10.1.130.1:1234", conf.Host)
-
-	// invalid k8s configuration
-	yamlStr2 := `
-apiserver:
-  k8s:
-    qps: "invalid"
-`
-	provider2, err := config.NewYAML(config.Source(strings.NewReader(yamlStr2)))
-	assert.NoError(t, err)
-	_, err = getK8sRestConfig(provider2)
-	assert.Error(t, err)
-}
-
 func TestGetYARPCConfig(t *testing.T) {
 	yamlConfStr := `
 apiserver:
