+++ go/cmd/controllermgr/BUILD.bazel
@@ -18,7 +18,9 @@ go_library(
         "@com_github_go_logr_zapr//:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
         "@io_k8s_client_go//kubernetes/scheme:go_default_library",
+        "@io_k8s_client_go//rest:go_default_library",
         "@io_k8s_sigs_controller_runtime//:go_default_library",
+        "@org_uber_go_config//:go_default_library",
         "@org_uber_go_fx//:go_default_library",
         "@org_uber_go_zap//:go_default_library",
     ],
