+++ go/cmd/controllermgr/config/base.yaml
@@ -4,8 +4,3 @@ controllermgr:
   leaderElection: false
   leaderElectionID: decaf1259.michelangelo.uber.com
   port: 9443
-
-controllers:
-  rayCluster:
-    k8sQps: 300
-    k8sBurst: 600
