+++ go/cmd/controllermgr/main.go
@@ -4,17 +4,24 @@ import (
 	"github.com/go-logr/zapr"
 	"go.uber.org/fx"
 	"go.uber.org/zap"
+	uberconfig "go.uber.org/config"
 	"k8s.io/apimachinery/pkg/runtime"
 	kubescheme "k8s.io/client-go/kubernetes/scheme"
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	"github.com/michelangelo-ai/michelangelo/go/base/config"
+	baseconfig"github.com/michelangelo-ai/michelangelo/go/base/config"
 	"github.com/michelangelo-ai/michelangelo/go/base/env"
 	"github.com/michelangelo-ai/michelangelo/go/base/zapfx"
 	"github.com/michelangelo-ai/michelangelo/go/components/ray"
 	"github.com/michelangelo-ai/michelangelo/go/components/spark"
 	"github.com/michelangelo-ai/michelangelo/go/controllermgr"
 	v2pb "github.com/michelangelo-ai/michelangelo/proto/api/v2"
+	"k8s.io/client-go/rest"
+)
+
+const (
+	k8sConfigKey           = "controllermgr.k8s"
+	metadataStorageConfigKey = "controllermgr.metadataStorage"
 )
 
 // scheme provides a Kubernetes runtime.Scheme object.
@@ -52,12 +59,23 @@ func options() fx.Option {
 		zapfx.Module,
 		fx.Provide(scheme),
 		spark.Module,
+		uberconfig.Module,
+		fx.Provide(getK8sConfig),
+		fx.Provide(getMetadataStorageConfig),
 		ray.Module,
 		controllermgr.Module,
 		fx.Invoke(func(logger *zap.Logger) {
 			ctrl.SetLogger(zapr.NewLogger(logger))
 		}),
 	)
+}
+
+func getK8sConfig(provider uberconfig.Provider) (*rest.Config, error) {
+	return baseconfig.GetK8sConfig(provider, k8sConfigKey)
+}
+
+func getMetadataStorageConfig(provider uberconfig.Provider) (*storage.MetadataStorageConfig, error) {
+	return baseconfig.GetMetadataStorageConfig(provider, metadataStorageConfigKey)
 }
 
 // main initializes and runs the application.
