+++ go/components/ray/cluster/module.go
@@ -6,24 +6,22 @@ import (
 
 	"github.com/michelangelo-ai/michelangelo/go/base/env"
 	rayv1 "github.com/ray-project/kuberay/ray-operator/pkg/client/clientset/versioned/typed/ray/v1"
+	"k8s.io/client-go/rest"
+
 )
 
 var (
 	// Module FX
 	Module = fx.Options(
-		fx.Provide(newConfig),
 		fx.Invoke(register),
 	)
 )
 
 func register(
-	conf Config,
+	restConfig *rest.Config,
 	env env.Context,
 	mgr manager.Manager,
 ) error {
-	restConfig := mgr.GetConfig()
-	restConfig.QPS = conf.QPS
-	restConfig.Burst = conf.Burst
 	rayClient, err := rayv1.NewForConfig(restConfig)
 	if err != nil {
 		return err
