+++ go/components/ray/job/BUILD.bazel
@@ -11,13 +11,13 @@ go_library(
     deps = [
         "//go/api/utils:go_default_library",
         "//go/base/env:go_default_library",
-        "//go/components/ray/cluster:go_default_library",
         "//proto/api/v2:go_default_library",
         "@com_github_go_logr_logr//:go_default_library",
         "@com_github_ray_project_kuberay_ray_operator//apis/ray/v1:go_default_library",
         "@com_github_ray_project_kuberay_ray_operator//pkg/client/clientset/versioned/typed/ray/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/types:go_default_library",
+        "@io_k8s_client_go//rest:go_default_library",
         "@io_k8s_sigs_controller_runtime//:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/client:go_default_library",
         "@io_k8s_sigs_controller_runtime//pkg/log:go_default_library",
