+++ go/components/ray/job/module.go
@@ -5,8 +5,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 
 	"github.com/michelangelo-ai/michelangelo/go/base/env"
-	"github.com/michelangelo-ai/michelangelo/go/components/ray/cluster"
 	rayv1 "github.com/ray-project/kuberay/ray-operator/pkg/client/clientset/versioned/typed/ray/v1"
+	"k8s.io/client-go/rest"
 )
 
 var (
@@ -17,13 +17,10 @@ var (
 )
 
 func register(
-	conf cluster.Config,
+	restConfig *rest.Config,
 	env env.Context,
 	mgr manager.Manager,
 ) error {
-	restConfig := mgr.GetConfig()
-	restConfig.QPS = conf.QPS
-	restConfig.Burst = conf.Burst
 	rayClient, err := rayv1.NewForConfig(restConfig)
 	if err != nil {
 		return err
